[utils] Move filter, serializeValue, and stringifyAsLabel to @base-ui/utils
