+++ packages/react/src/utils/resolveValueLabel.tsx
@@ -1,6 +1,7 @@
 'use client';
 import * as React from 'react';
-import { serializeValue } from './serializeValue';
+import { serializeValue } from '@base-ui/utils/serializeValue';
+import { stringifyAsLabel } from '@base-ui/utils/stringifyAsLabel';
 
 type ItemRecord = Record<string, React.ReactNode>;
 type ItemsInput = ItemRecord | ReadonlyArray<LabeledItem> | ReadonlyArray<Group<any>> | undefined;
@@ -57,20 +58,7 @@ export function hasNullItemLabel(items: ItemsInput): boolean {
   return false;
 }
 
-export function stringifyAsLabel(item: any, itemToStringLabel?: (item: any) => string) {
-  if (itemToStringLabel && item != null) {
-    return itemToStringLabel(item) ?? '';
-  }
-  if (item && typeof item === 'object') {
-    if ('label' in item && item.label != null) {
-      return String(item.label);
-    }
-    if ('value' in item) {
-      return String(item.value);
-    }
-  }
-  return serializeValue(item);
-}
+export { stringifyAsLabel } from '@base-ui/utils/stringifyAsLabel';
 
 export function stringifyAsValue(item: any, itemToStringValue?: (item: any) => string) {
   if (itemToStringValue && item != null) {
