+++ packages/utils/src/stringifyAsLabel.ts
@@ -1,0 +1,16 @@
+import { serializeValue } from './serializeValue';
+
+export function stringifyAsLabel(item: any, itemToStringLabel?: (item: any) => string) {
+  if (itemToStringLabel && item != null) {
+    return itemToStringLabel(item) ?? '';
+  }
+  if (item && typeof item === 'object') {
+    if ('label' in item && item.label != null) {
+      return String(item.label);
+    }
+    if ('value' in item) {
+      return String(item.value);
+    }
+  }
+  return serializeValue(item);
+}
